Fix layout issues on large screeens
