+++ intelligent_remote_academia-frontend/src/components/constants/layoutConsts.js
@@ -1,0 +1,45 @@
+import { makeStyles, useTheme } from "@material-ui/core/styles";
+
+const drawerWidth = 240;
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    display: "flex",
+  },
+  drawer: {
+    [theme.breakpoints.up("sm")]: {
+      width: drawerWidth,
+      flexShrink: 0,
+    },
+  },
+  appBar: {
+    [theme.breakpoints.up("sm")]: {
+      width: `calc(100% - ${drawerWidth}px)`,
+      marginLeft: drawerWidth,
+    },
+  },
+  menuButton: {
+    marginRight: theme.spacing(2),
+    [theme.breakpoints.up("sm")]: {
+      display: "none",
+    },
+  },
+  // necessary for content to be below app bar
+  toolbar: theme.mixins.toolbar,
+  drawerPaper: {
+    width: drawerWidth,
+  },
+  content: {
+    marginTop: "70px",
+    flexGrow: 1,
+    padding: theme.spacing(3),
+  },
+  active: {
+    backgroundColor: "#E1F0FF",
+  },
+  title: {
+    flexGrow: 1,
+  },
+}));
+
+export { useStyles, drawerWidth };
