--- intelligent_remote_academia-frontend/src/components/constants/mainContentConsts.js
@@ -1,14 +1,0 @@
-import { makeStyles } from "@material-ui/core/styles";
-
-const useStyles = makeStyles({
-  container: {
-    marginTop: "40px",
-    padding: "40px 20px",
-  },
-
-  menu: {
-    width: "1500px",
-  },
-});
-
-export { useStyles };
