+++ intelligent_remote_academia-frontend/src/components/constants/navbarConsts.js
@@ -1,6 +1,3 @@
-import { makeStyles } from "@material-ui/core/styles";
-import { drawerWidth } from "./sidebarConsts";
-
 const accountMenu = [
   // {
   //   text: "Profile",
@@ -14,25 +11,4 @@ const accountMenu = [
 
 const brandName = "Intelligent Remote Academia";
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-  },
-  appBar: {
-    [theme.breakpoints.up("sm")]: {
-      width: `calc(100% - ${drawerWidth}px)`,
-      marginLeft: drawerWidth,
-    },
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-    [theme.breakpoints.up("sm")]: {
-      display: "none",
-    },
-  },
-  title: {
-    flexGrow: 1,
-  },
-}));
-
-export { accountMenu, brandName, useStyles };
+export { accountMenu, brandName };
