+++ intelligent_remote_academia-frontend/src/components/constants/sidebarConsts.js
@@ -1,18 +1,16 @@
-import { makeStyles } from '@material-ui/core/styles';
-
 //Sidebar link constatnts
 const sideBarLinks = [
   {
-    text: 'Subjects',
-    url: '/subjects',
+    text: "Subjects",
+    url: "/subjects",
   },
   {
-    text: 'Attendance',
-    url: '/attendance',
+    text: "Attendance",
+    url: "/attendance",
   },
   {
-    text: 'Announcements',
-    url: '/announcements',
+    text: "Announcements",
+    url: "/announcements",
   },
   {
     text: "Fee Challan Form",
@@ -20,30 +18,4 @@ const sideBarLinks = [
   },
 ];
 
-//Sidebar Css constants
-const drawerWidth = 240;
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    display: 'flex',
-  },
-  drawer: {
-    [theme.breakpoints.up('sm')]: {
-      width: drawerWidth,
-      flexShrink: 0,
-    },
-  },
-  toolbar: theme.mixins.toolbar,
-  drawerPaper: {
-    width: drawerWidth,
-  },
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing(3),
-  },
-  active: {
-    backgroundColor: "#E1F0FF",
-  },
-}));
-
-export { useStyles, sideBarLinks, drawerWidth };
+export { sideBarLinks };
